// tools/jmespath-tester/client/types.ts
export type JmespathTesterMessages = Readonly<{
  meta: {
    name: string
    description: string
  }
  jsonLabel: string
  jsonDescription: string
  jsonPlaceholder: string
  queryLabel: string
  queryDescription: string
  queryPlaceholder: string
  examplesLabel: string
  examplesPlaceholder: string
  formatJsonLabel: string
  importFromFileLabel: string
  useSampleLabel: string
  clearLabel: string
  copyQueryLabel: string
  copyResultLabel: string
  copiedLabel: string
  resultLabel: string
  resultDescription: string
  resultCountLabel: string
  resultEmptyDescription: string
  noResultsLabel: string
  downloadJsonLabel: string
  invalidJsonLabel: string
  invalidQueryLabel: string
  exampleLastNamesLabel: string
  exampleAdultsLabel: string
  exampleOrdersLabel: string
}>